\docType{data}
\name{NR.C66.trr}
\alias{NR.C66.trr}
\title{Grand-average ERP data from 66 participants in auditory oddball task.}
\format{A 200 x 31 matrix. Rows are time samples (in increments of 5 ms).
  Columns are electrode sites.}
\source{
  http://psychophysiology.cpmc.columbia.edu/software/CSDtoolbox/CSDtoolbox.zip
}
\description{
  Grand-average ERP data derived from 66 healthy adults
  that participated in an auditory oddball task (Kayser &
  Tenke, 2006a). The data refers to their complex tones
  condition, where participants were instructed to press
  the right button when rare stimuli were perceived. The
  data was acquired with a 31 electrode montage, at a
  sampling frequency of 200 Hz.  This dataset has 200
  samples, ranging from time 0 to 995 ms post-stimulus
  onset, and has been referenced to the tip of the nose.
}
\references{
  Kayser, J., Tenke, C.E. (2006a). Principal components
  analysis of Laplacian waveforms as a generic method for
  identifying ERP generator patterns: I. Evaluation with
  auditory oddball tasks. Clinical Neurophysiology, 117(2),
  348-368.
}
\keyword{datasets}

